refactor: update manifest handling and restructure project files
