+++ .vscode/launch.json
@@ -10,7 +10,6 @@
       "name": "Launch Chrome against example-extension-usage",
       "url": "http://localhost:49675/",
       "webRoot": "${workspaceFolder}/dist",
-      "pathMapping": { "__webpack_modules__..": "${workspaceFolder}" },
       "runtimeArgs": ["--load-extension=${workspaceFolder}/dist", "--auto-open-devtools-for-tabs"],
       "timeout": 100000,
       "includeDefaultArgs": true,
@@ -18,9 +17,6 @@
       "internalConsoleOptions": "openOnSessionStart",
       "userDataDir": "${workspaceRoot}/.vscode/chrome",
       "diagnosticLogging": true,
-      "sourceMapPathOverrides": {
-        "webpack:///./*": "${webRoot}/*"
-      },
       "preLaunchTask": "npm: example-extension-usage",
     },
     {
@@ -29,17 +25,13 @@
       "name": "Launch Chrome against slovensko.sk",
       "url": "https://slovensko.sk/",
       "webRoot": "${workspaceFolder}/dist",
-      "pathMapping": { "__webpack_modules__..": "${workspaceFolder}" },
       "runtimeArgs": ["--load-extension=${workspaceFolder}/dist"],
       "timeout": 100000,
       "includeDefaultArgs": true,
       "trace": true,
       "internalConsoleOptions": "openOnSessionStart",
       "userDataDir": "${workspaceRoot}/.vscode/chrome",
-      "diagnosticLogging": true,
-      "sourceMapPathOverrides": {
-        "webpack:///./*": "${webRoot}/*"
-      }
+      "diagnosticLogging": true
     },
     {
       "type": "firefox",
@@ -52,13 +44,7 @@
           "PathConversion": "Debug",
           "default": "Error"
         }
-      },
-      "pathMappings": [
-        {
-          "url": "webpack://autogram-extension/src",
-          "path": "${workspaceFolder}/src"
-        }
-      ]
+      }
     },
     {
       "type": "firefox",
@@ -78,14 +64,8 @@
           "PathConversion": "Debug",
           "default": "Error"
         }
-      },
-      "pathMappings": [
-        {
-          "url": "webpack://autogram-extension/src",
-          "path": "${workspaceFolder}/src"
-        }
-      ]
+      }
     }
-    
+
   ]
 }
