+++ autogram-extension/manifest/manifest-types.ts
